Created Tagger init code and interface.
Similar to UIMA's sandbox implementation of an HMM tagger, I moved all
init code to the Annotator's initialize() method. The taggers are now
initialized by model files that exist in the project as resources. This
location may change in the future, but right now it's
org.werti.resources. The HMM tagger now works splendidly, and is very
fast. I have to test its correctness, though, as it is only a toy
implementation.

The existance of more than one tagger also implies that we'll have to
deal with more than one tag set. A parameter config for all PoS-related
tagging and/or a translator that reads tags in and standardizes over
them might be needed.

I could still try to implement the TreeTagger as well. Or the one the
thing Detmar talked about. What was its name again? linkpipe?