Changed Interface.
The interface is now a JSP page, which should be easier to maintain.
Changed Interface.java so it doesn't output anything, but process
the parameters of user's requests. Currently it only logs those to the
main log and does nothing (so the system overall doesn't work).

It would be nice if Interface.java could actually create an Analysis
Engine Descriptor file and autonomously import appropriate TAEs and
create a component work flow.

Damn those corporate buzz-words. I'm off. Good night.