Modularization. WERTiContext. Properties. Langs.
A lot of work on modularizing everything and making it more portable to
ther systems. That'll teach me to do this right from the start next
time.

Using .properties files now. Migration not! complete, need to migrate
the analysis engines, too, so they use the context.

Added benefit: we might get i18n support for the analysis process
easily this way.